server/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
cryptography==42.0.1
urllib3==2.1.0
python-dotenv==1.0.1
requests==2.31.0
